Add CrtCodePage Procedure.
The CrtCodePage Procedure allows the user to set a specific code page for the windows console.

CrtCodePage(0) will set the console to the windows code page reported by GetACP.  This default happens when CRT is initialized.

CrtCodePage(-1) reverts back to the code page of the console before CRT was initialized

Any other value sets the codepage to the one specified.

by default:
The console code page is only saved once, during initialization

The Codepage is only set once during initialiazation, and default is to the system codepage returned by GetACP

The Codepage is only restored once during finialization