Add Procedures and Flags to control CodePage switching and Compatibility on Windows
Added new Global Variable:  OriginalConsoleOutputCP
Used to Store the Original Console Code Page so it can be restored in Finalization.  This is useful when the program is run from a command line so the console is always restored to the state it was in before the program was run.

Added 2 Boolean Variables:

TP_Compatible: Defult to False.  When True the Console Original Code Page will be used, when False, the System Codepage from GetACP will be transfered to the Console.

Switch_CodePage:  Default to True. 

When Switch_Codepage is True and TP_Compatible is False,  the console code page will be switched every read and write with the system codepage from GetACP as the previous CRT Version does.  
When Switch_Codepage is True and TP_Compatible is True,  the console code page will be reset to the original after every read and write.

Upon Initialiaztion Variable Defaults are Set.

Added 2 Procedures:

CRT_Switch_CodePage:
Takes a Boolean parameter, Switch_codepage is set to the parameter, and Codepages are set as needed to be in the correct mode of operation after the procedure completes.

CRT_TP_Compatible:
Takes a Boolean Parameter.  TP_Compatible is set to the Parameter and CodePage is set as needed to operatin in the correct mode after the procedure completes.

CRTRead and CRTWrite were modified to use the control variables and Switch and restore the codepages as needed.