Fix incorrect type for CtrlKeyState in Function RemapScanCode
CtrlKeyState is passed from KeyEvent.dwControlKeyState and was originally a Byte, however according to the documenation of dwControlkeyState here:
https://docs.microsoft.com/en-us/windows/console/key-event-record-str

dwControlKeyState is a word variable. 
I encountered a range check error when pushing the right arrow key and found that the dwControlKeyState being received was $0120 thus putting it out of range of a byte.

with CtrlKeyState propertly defined as a Word, the function now opertes correctly and does not produce a range check error, and also returns the correct remapping of the key.